requirements.txt
discord.py==1.5.1
canvasapi==2.0.0
discord==1.0.1
python-dotenv==0.14.0
